- mpqs_factor() definition changed to match mpqs.c ver 1.9 (1.4 -last stable).
git-svn-id: https://ggnfs.svn.sourceforge.net/svnroot/ggnfs/trunk@142 80ef69ca-9341-0410-9d60-8acc7774fa31

+++ src/lasieve4/lasieve.h
@@ -74,6 +74,6 @@ uint32_t pr32_seek(pr32_struct * ps, uint32_t lb);
 int psp(mpz_t);
 
 /* mpqs.c */
-int mpqs_factor(mpz_t N, size_t max_bits, mpz_t **factors);
+long mpqs_factor(mpz_t N, long max_bits, mpz_t **factors);
 
 #endif
